+++ CHANGES.md
@@ -1,10 +1,16 @@
 Bridgestyle Changelog
 =====================
 
-## 0.1.4 (unreleased)
+## 0.1.5 (unreleased)
 ---------------------
 
 - Nothing changed yet.
+
+
+## 0.1.4 (2025-08-26)
+---------------------
+- ArcGIS: Picture fill support, improved hatch fills, color substitution, marker along line (thanks @aaime) 
+- Fix URLs in README for PyPI (thanks @geraldo)
 
 
 ## 0.1.3 (2025-06-24)